+++ gradle.properties
@@ -12,7 +12,7 @@ makeevrserg.java.ktarget=21
 # Project
 makeevrserg.project.name=SoulKeeper
 makeevrserg.project.group=ru.astrainteractive.soulkeeper
-makeevrserg.project.version.string=0.2.2
+makeevrserg.project.version.string=0.2.3
 makeevrserg.project.description=Keep your items after death
 makeevrserg.project.developers=makeevrserg|Makeev Roman|makeevrserg@gmail.com
 makeevrserg.project.url=https://github.com/Astra-Interactive/SoulKeeper
