--- modules/core/src/main/kotlin/ru/astrainteractive/soulkeeper/core/util/KrateExt.kt
@@ -1,8 +1,0 @@
-package ru.astrainteractive.soulkeeper.core.util
-
-import ru.astrainteractive.klibs.kstorage.api.cache.CacheOwner
-import kotlin.reflect.KProperty
-
-operator fun <T> CacheOwner<T>.getValue(thisRef: Any, property: KProperty<*>): T {
-    return this.cachedValue
-}
