+++ modules/dao/src/main/kotlin/ru/astrainteractive/soulkeeper/module/souls/database/table/SoulTable.kt
