+++ modules/dao/src/main/kotlin/ru/astrainteractive/soulkeeper/module/souls/io/SoulFile.kt
@@ -1,0 +1,17 @@
+package ru.astrainteractive.soulkeeper.module.souls.io
+
+import ru.astrainteractive.soulkeeper.module.souls.io.model.BukkitSoul
+import ru.astrainteractive.soulkeeper.module.souls.io.model.Soul
+
+internal interface SoulFile {
+    fun read(soul: Soul): Result<BukkitSoul>
+
+    fun write(soul: BukkitSoul): Result<Unit>
+
+    fun delete(soul: Soul)
+
+    companion object {
+        val Soul.fileNameWithoutExtension: String
+            get() = "${ownerUUID}_${createdAt.toEpochMilli()}"
+    }
+}
