+++ modules/dao/src/main/kotlin/ru/astrainteractive/soulkeeper/module/souls/io/model/Soul.kt
@@ -1,0 +1,15 @@
+package ru.astrainteractive.soulkeeper.module.souls.io.model
+
+import org.bukkit.Location
+import java.time.Instant
+import java.util.UUID
+
+interface Soul {
+    val ownerUUID: UUID
+    val ownerLastName: String
+    val createdAt: Instant
+    val isFree: Boolean
+    val location: Location
+    val hasItems: Boolean
+    val hasXp: Boolean
+}
