--- modules/database/build.gradle.kts
@@ -1,28 +1,0 @@
-plugins {
-    kotlin("jvm")
-    kotlin("plugin.serialization")
-}
-
-dependencies {
-    // Kotlin
-    implementation(libs.bundles.kotlin)
-    implementation(libs.bundles.exposed)
-    // Bukkit
-    compileOnly(libs.minecraft.paper.api)
-    // AstraLibs
-    implementation(libs.minecraft.astralibs.core)
-    implementation(libs.minecraft.astralibs.exposed)
-    implementation(libs.minecraft.astralibs.command)
-    implementation(libs.minecraft.astralibs.command.bukkit)
-    implementation(libs.minecraft.astralibs.menu.bukkit)
-    implementation(libs.minecraft.astralibs.core.bukkit)
-    // klibs
-    implementation(libs.klibs.kstorage)
-    implementation(libs.klibs.mikro.core)
-    compileOnly(libs.minecraft.packetevents)
-    // Test
-    testImplementation(libs.bundles.testing.kotlin)
-    testImplementation(libs.tests.kotlin.test)
-    // Local
-    implementation(projects.modules.core)
-}
