+++ modules/worker/build.gradle.kts
@@ -25,5 +25,5 @@ dependencies {
     testImplementation(libs.tests.kotlin.test)
     // Local
     implementation(projects.modules.core)
-    implementation(projects.modules.database)
+    implementation(projects.modules.dao)
 }
