+++ modules/worker/src/main/kotlin/ru/astrainteractive/soulkeeper/module/souls/domain/GetNearestSoulUseCase.kt
@@ -3,7 +3,7 @@ package ru.astrainteractive.soulkeeper.module.souls.domain
 import org.bukkit.entity.Player
 import ru.astrainteractive.astralibs.logging.JUtiltLogger
 import ru.astrainteractive.astralibs.logging.Logger
-import ru.astrainteractive.soulkeeper.module.souls.database.dao.SoulsDao
+import ru.astrainteractive.soulkeeper.module.souls.dao.SoulsDao
 import ru.astrainteractive.soulkeeper.module.souls.database.model.DatabaseSoul
 
 /**
