+++ modules/worker/src/main/kotlin/ru/astrainteractive/soulkeeper/module/souls/domain/PickUpExpUseCase.kt
@@ -5,8 +5,8 @@ import ru.astrainteractive.astralibs.logging.JUtiltLogger
 import ru.astrainteractive.astralibs.logging.Logger
 import ru.astrainteractive.soulkeeper.core.plugin.SoulsConfig
 import ru.astrainteractive.soulkeeper.core.util.playSound
-import ru.astrainteractive.soulkeeper.module.souls.database.dao.SoulsDao
-import ru.astrainteractive.soulkeeper.module.souls.database.model.BukkitSoul
+import ru.astrainteractive.soulkeeper.module.souls.dao.SoulsDao
+import ru.astrainteractive.soulkeeper.module.souls.io.model.BukkitSoul
 
 internal class PickUpExpUseCase(
     private val collectXpSoundProvider: () -> SoulsConfig.Sounds.SoundConfig,
@@ -24,7 +24,7 @@ internal class PickUpExpUseCase(
         bukkitSoul.location.playSound(collectXpSoundProvider.invoke())
         player.giveExp(bukkitSoul.exp)
         soulsDao.updateSoul(
-            bukkitSoul.copy(
+            soul = bukkitSoul.copy(
                 exp = 0,
             )
         )
