+++ modules/worker/src/main/kotlin/ru/astrainteractive/soulkeeper/module/souls/domain/PickUpSoulUseCase.kt
@@ -8,8 +8,8 @@ import ru.astrainteractive.klibs.mikro.core.dispatchers.KotlinDispatchers
 import ru.astrainteractive.soulkeeper.core.plugin.SoulsConfig
 import ru.astrainteractive.soulkeeper.core.util.playSound
 import ru.astrainteractive.soulkeeper.core.util.spawnParticle
-import ru.astrainteractive.soulkeeper.module.souls.database.dao.SoulsDao
-import ru.astrainteractive.soulkeeper.module.souls.database.model.BukkitSoul
+import ru.astrainteractive.soulkeeper.module.souls.dao.SoulsDao
+import ru.astrainteractive.soulkeeper.module.souls.io.model.BukkitSoul
 
 @Suppress("LongParameterList")
 internal class PickUpSoulUseCase(
