+++ modules/worker/src/main/kotlin/ru/astrainteractive/soulkeeper/module/souls/domain/armorstand/ShowArmorStandUseCase.kt
@@ -1,7 +1,7 @@
 package ru.astrainteractive.soulkeeper.module.souls.domain.armorstand
 
 import org.bukkit.entity.Player
-import ru.astrainteractive.soulkeeper.module.souls.database.model.Soul
+import ru.astrainteractive.soulkeeper.module.souls.io.model.Soul
 
 internal interface ShowArmorStandUseCase {
     fun destroy(player: Player, ids: Collection<Int>)
