+++ plugin/src/main/kotlin/ru/astrainteractive/soulkeeper/SoulKeeper.kt
@@ -1,8 +1,8 @@
 package ru.astrainteractive.soulkeeper
 
+import ru.astrainteractive.astralibs.lifecycle.LifecyclePlugin
 import ru.astrainteractive.astralibs.logging.JUtiltLogger
 import ru.astrainteractive.astralibs.logging.Logger
-import ru.astrainteractive.soulkeeper.core.plugin.LifecyclePlugin
 import ru.astrainteractive.soulkeeper.di.RootModule
 
 class SoulKeeper :
