+++ plugin/src/main/kotlin/ru/astrainteractive/soulkeeper/command/SoulsCommandRegistry.kt
@@ -13,18 +13,18 @@ import ru.astrainteractive.astralibs.command.api.parser.CommandParser
 import ru.astrainteractive.astralibs.command.api.util.PluginExt.registerCommand
 import ru.astrainteractive.astralibs.kyori.KyoriComponentSerializer
 import ru.astrainteractive.astralibs.permission.BukkitPermissibleExt.toPermissible
+import ru.astrainteractive.astralibs.util.clickable
+import ru.astrainteractive.astralibs.util.isEmpty
+import ru.astrainteractive.astralibs.util.orEmpty
 import ru.astrainteractive.klibs.kstorage.api.Krate
+import ru.astrainteractive.klibs.kstorage.util.getValue
+import ru.astrainteractive.soulkeeper.core.datetime.TimeAgoFormatter
+import ru.astrainteractive.soulkeeper.core.datetime.TimeAgoTranslationFormatter
 import ru.astrainteractive.soulkeeper.core.plugin.PluginPermission
 import ru.astrainteractive.soulkeeper.core.plugin.PluginTranslation
-import ru.astrainteractive.soulkeeper.core.util.clickable
-import ru.astrainteractive.soulkeeper.core.util.getValue
-import ru.astrainteractive.soulkeeper.core.util.isEmpty
-import ru.astrainteractive.soulkeeper.core.util.orEmpty
-import ru.astrainteractive.soulkeeper.datetime.TimeAgoFormatter
-import ru.astrainteractive.soulkeeper.datetime.TimeAgoTranslationFormatter
-import ru.astrainteractive.soulkeeper.module.souls.database.dao.SoulsDao
+import ru.astrainteractive.soulkeeper.module.souls.dao.SoulsDao
 import ru.astrainteractive.soulkeeper.module.souls.database.model.DatabaseSoul
-import ru.astrainteractive.soulkeeper.module.souls.database.model.Soul
+import ru.astrainteractive.soulkeeper.module.souls.io.model.Soul
 import ru.astrainteractive.soulkeeper.module.souls.worker.call.SoulCallRenderer
 
 internal class SoulsCommandRegistry(
